+++ vec3.h
@@ -5,36 +5,36 @@
 class vec3 {
 public:
     vec3() {}
-    vec3(double e0, double e1, double e2) {e[0] = e0; e[1] = e1; e[2] = e2;}
+    vec3(float e0, float e1, float e2) {e[0] = e0; e[1] = e1; e[2] = e2;}
 
-    inline double x() const {return e[0];}
-    inline double y() const {return e[1];}
-    inline double z() const {return e[2];}
-    inline double r() const {return e[0];}
-    inline double g() const {return e[1];}
-    inline double b() const {return e[2];}
+    inline float x() const {return e[0];}
+    inline float y() const {return e[1];}
+    inline float z() const {return e[2];}
+    inline float r() const {return e[0];}
+    inline float g() const {return e[1];}
+    inline float b() const {return e[2];}
 
     inline const vec3& operator+() const {return *this;}
     inline vec3 operator-() const {return vec3(-e[0], -e[1], -e[2]);}
-    inline double operator[](int i) const {return e[i];}
-    inline double& operator[](int i) {return e[i];}
+    inline float operator[](int i) const {return e[i];}
+    inline float& operator[](int i) {return e[i];}
 
     inline vec3& operator+=(const vec3& v2);
     inline vec3& operator-=(const vec3& v2);
     inline vec3& operator*=(const vec3& v2);
     inline vec3& operator/=(const vec3& v2);
-    inline vec3& operator*=(const double t);
-    inline vec3& operator/=(const double t);
+    inline vec3& operator*=(const float t);
+    inline vec3& operator/=(const float t);
 
-    inline double length() const {
+    inline float length() const {
       return sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
     }
-    inline double squaredLength() const {
-      return e[0]*e[0] + e[1]*e[1] + e[2]*e[2];
+    inline float squaredLength() const {
+      return (e[0]*e[0]) + (e[1]*e[1]) + (e[2]*e[2]);
     }
     inline void makeUnitVector();
 
-    double e[3];
+    float e[3];
 };
 
 inline std::istream& operator>>(std::istream &is, vec3 &t) {
@@ -48,7 +48,7 @@ inline std::ostream& operator<<(std::ostream &os, const vec3 &t) {
 }
 
 inline void vec3::makeUnitVector() {
-  double k = 1.0 / sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
+  float k = 1.0 / sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
   e[0] *= k;
   e[1] *= k;
   e[2] *= k;
@@ -70,19 +70,19 @@ inline vec3 operator/(const vec3 &v1, const vec3 &v2) {
   return vec3(v1.e[0] / v2.e[0], v1.e[1] / v2.e[1], v1.e[2] / v2.e[2]);
 }
 
-inline vec3 operator*(double t, const vec3 &v) {
+inline vec3 operator*(float t, const vec3 &v) {
   return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
-inline vec3 operator*(const vec3 &v, double t) {
+inline vec3 operator*(const vec3 &v, float t) {
   return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
-inline vec3 operator/(const vec3 &v, double t) {
+inline vec3 operator/(const vec3 &v, float t) {
   return vec3(v.e[0]/t, v.e[1]/t, v.e[2]/t);
 }
 
-inline double dot(const vec3 &v1, const vec3 &v2) {
+inline float dot(const vec3 &v1, const vec3 &v2) {
   return v1.e[0] * v2.e[0] + v1.e[1] * v2.e[1] + v1.e[2] * v2.e[2];
 }
 
@@ -120,15 +120,15 @@ inline vec3& vec3::operator/=(const vec3 &v) {
   return *this;
 }
 
-inline vec3& vec3::operator*=(const double t) {
+inline vec3& vec3::operator*=(const float t) {
   e[0] *= t;
   e[1] *= t;
   e[2] *= t;
   return *this;
 }
 
-inline vec3& vec3::operator/=(const double t) {
-  double k = 1.0/t;
+inline vec3& vec3::operator/=(const float t) {
+  float k = 1.0/t;
   e[0] *= k;
   e[1] *= k;
   e[2] *= k;
